requirements.txt
python-docx==0.8.11
openai==1.3.5
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.1
httpx==0.24.1
anyio==3.7.1
pydantic==2.4.2
pydantic-core==2.10.1
typing-extensions==4.7.1
distro==1.8.0
tqdm==4.65.0
certifi==2023.7.22
idna==3.4
sniffio==1.3.0
h11==0.14.0
httpcore==0.17.3
olefile==0.46
